Move session request params handling in an http listener
