+++ src/Glpi/Http/Listener/SessionVariables.php
@@ -32,41 +32,48 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Config\LegacyConfigurators;
+namespace Glpi\Http\Listener;
 
-use Glpi\Config\ConfigProviderHasRequestTrait;
-use Glpi\Config\ConfigProviderWithRequestInterface;
-use Glpi\Config\LegacyConfigProviderInterface;
+use Glpi\Kernel\ListenersPriority;
 use Glpi\Toolbox\URL;
 use Session;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
+use Symfony\Component\HttpKernel\KernelEvents;
 
-final class SessionConfig implements LegacyConfigProviderInterface, ConfigProviderWithRequestInterface
+final class SessionVariables implements EventSubscriberInterface
 {
-    use ConfigProviderHasRequestTrait;
-
-    public function execute(): void
+    public static function getSubscribedEvents(): array
     {
-        if (!isset($_SESSION["MESSAGE_AFTER_REDIRECT"])) {
-            $_SESSION["MESSAGE_AFTER_REDIRECT"] = [];
+        return [
+            KernelEvents::REQUEST => ['onKernelRequest', ListenersPriority::REQUEST_LISTENERS_PRIORITIES[self::class]],
+        ];
+    }
+
+    public function onKernelRequest(RequestEvent $event): void
+    {
+        if (!$event->isMainRequest()) {
+            return;
         }
 
-        $request = $this->getRequest();
+        $request = $event->getRequest();
 
         // Manage force tab
         if ($request->query->has('forcetab')) {
             $itemtype = URL::extractItemtypeFromUrlPath($request->getPathInfo());
             if ($itemtype !== null) {
-                Session::setActiveTab($itemtype, $_REQUEST['forcetab']);
+                Session::setActiveTab($itemtype, $request->get('forcetab'));
             }
         }
 
         // Manage tabs
-        if (isset($_REQUEST['glpi_tab']) && isset($_REQUEST['itemtype'])) {
-            Session::setActiveTab($_REQUEST['itemtype'], $_REQUEST['glpi_tab']);
+        if ($request->get('glpi_tab') && $request->get('itemtype')) {
+            Session::setActiveTab($request->get('itemtype'), $request->get('glpi_tab'));
         }
+
         // Override list-limit if choosen
-        if (isset($_REQUEST['glpilist_limit'])) {
-            $_SESSION['glpilist_limit'] = $_REQUEST['glpilist_limit'];
+        if ($request->get('glpilist_limit')) {
+            $_SESSION['glpilist_limit'] = $request->get('glpilist_limit');
         }
     }
 }
